requirements.txt
asgiref==3.5.2
autopep8==2.0.0
azure-common==1.1.28
azure-core==1.26.1
azure-keyvault-secrets==4.6.0
azure-mgmt-core==1.3.2
azure-mgmt-storage==21.0.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-storages-azure==1.6.8
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.7.1
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
